Fix implicit conversion isssue

+++ arcade/examples/transform_feedback.py
@@ -43,9 +43,9 @@ class MyGame(arcade.Window):
             void main() {
                 // Let's just give them a "random" color based on the vertex id
                 color = vec3(
-                    mod((gl_VertexID * 100 % 11) / 10.0, 1.0),
-                    mod((gl_VertexID * 100 % 27) / 10.0, 1.0),
-                    mod((gl_VertexID * 100 % 71) / 10.0, 1.0));
+                    mod(float(gl_VertexID * 100 % 11) / 10.0, 1.0),
+                    mod(float(gl_VertexID * 100 % 27) / 10.0, 1.0),
+                    mod(float(gl_VertexID * 100 % 71) / 10.0, 1.0));
                 // Pass the point position to primitive assembly
                 gl_Position = vec4(in_pos, 0.0, 1.0);
             }
